Introduced CoreData to store data after first launch

+++ TestTask/TestTask/String+Extension.swift
@@ -18,8 +18,10 @@ extension String {
     
     var string = ""
     for _ in 0..<length {
-      //Convert CharacterView back to String
-      string += String(charactersArray[Int(arc4random()) % charactersArray.count])
+      // The Int type is a 32-bit integer on the iPhone 5 and a 64-bit integer on the iPhone6.
+      // arc4random() returns a UInt32, which has twice the positive range of an Int on a 32 bit device.
+      // that means half of the time the number is too big for the cast.
+      string += String(charactersArray[Int(arc4random() % UInt32(charactersArray.count))])
     }
     
     return string
